Filter files in primary
