+++ src/test/java/com/artipie/rpm/meta/XmlEventPrimaryTest.java
@@ -21,7 +21,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.xmlunit.matchers.CompareMatcher;
 
 /**
- * Test for {@link XmlEvent.Primary}.
+ * Test for {@link XmlEventPrimary}.
  * @since 1.5
  */
 @SuppressWarnings("PMD.AvoidDuplicateLiterals")
@@ -46,7 +46,7 @@ class XmlEventPrimaryTest {
         writer.add(events.createStartElement("", "", "metadata"));
         writer.add(events.createNamespace("http://linux.duke.edu/metadata/common"));
         writer.add(events.createNamespace("rpm", "http://linux.duke.edu/metadata/rpm"));
-        new XmlEvent.Primary().add(
+        new XmlEventPrimary().add(
             writer,
             new FilePackage.Headers(new FilePackageHeader(file).header(), file, Digest.SHA256)
         );
@@ -60,7 +60,6 @@ class XmlEventPrimaryTest {
                 .ignoreWhitespace()
                 .ignoreElementContentWhitespace()
                 .normalizeWhitespace()
-                .withNodeFilter(node -> !"file".equals(node.getLocalName()))
         );
     }
