+++ src/test/resources-binary/AstoMetadataAddTest/primary-res.xml
@@ -95,9 +95,6 @@
                 <rpm:entry name="libgcc_s.so.1"/>
                 <rpm:entry name="libgcc_s.so.1(GCC_3.5)"/>
             </rpm:requires>
-            <file>/usr/lib/libdeflt.so.1.0</file>
-            <file type="dir">/usr/share/licenses/libdeflt1_0</file>
-            <file>/usr/share/licenses/libdeflt1_0/CDDL.Schily.txt</file>
         </format>
     </package>
     <package type="rpm">
@@ -143,12 +140,6 @@
                 <rpm:entry name="rtld(GNU_HASH)"/>
             </rpm:requires>
             <file>/usr/bin/abc</file>
-            <file type="dir">/usr/lib/.build-id/cb</file>
-            <file>/usr/lib/.build-id/cb/263ca8bdb2d581b1a12e604c0e30635e69c889</file>
-            <file type="dir">/usr/share/doc/abc</file>
-            <file>/usr/share/doc/abc/README.md</file>
-            <file>/usr/share/doc/abc/readmeaig</file>
-            <file>/usr/share/man/man1/abc.1.gz</file>
         </format>
     </package>
 </metadata>
